We need to implement each class on its own and a base activity for implementing UI stuff does not help us much. The menu functions are meant for us to override and allows us to define the implementation carefully for each to enhance the user experience.
