+++ app/build.gradle
@@ -29,14 +29,13 @@ android {
 dependencies {
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation "androidx.core:core-ktx:1.6.0"
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.0'
-    implementation 'com.airbnb.android:lottie:4.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
+    implementation 'com.airbnb.android:lottie:4.2.0'
     implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     // Coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2-native-mt'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2-native-mt'
 
     // Retrofit
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
