+++ app/src/main/java/org/musicbrainz/picard/barcodescanner/activities/AboutActivity.kt
@@ -19,15 +19,21 @@
  */
 package org.musicbrainz.picard.barcodescanner.activities
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.text.Html
 import android.text.method.LinkMovementMethod
+import android.view.Menu
+import android.view.MenuItem
 import android.view.View
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
 import org.musicbrainz.picard.barcodescanner.R
 import org.musicbrainz.picard.barcodescanner.databinding.ActivityAboutBinding
+import org.musicbrainz.picard.barcodescanner.util.Constants
 
-class AboutActivity : BaseActivity() {
+class AboutActivity : AppCompatActivity() {
     private lateinit var binding: ActivityAboutBinding
 
     private val contentMappings = mapOf(
@@ -50,6 +56,31 @@ class AboutActivity : BaseActivity() {
             val infoTextView = findViewById<View>(rView) as TextView
             infoTextView.text = Html.fromHtml(getString(rText), Html.FROM_HTML_MODE_LEGACY)
             infoTextView.movementMethod = LinkMovementMethod.getInstance()
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        // Inflate the menu items for use in the action bar
+        val inflater = menuInflater
+        inflater.inflate(R.menu.main_activity_actions, menu)
+        menu.findItem(R.id.action_about).isVisible = false
+        return super.onCreateOptionsMenu(menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle presses on the action bar items
+        return when (item.itemId) {
+            R.id.action_settings -> {
+                val preferencesIntent = Intent(this, PreferencesActivity::class.java)
+                startActivity(preferencesIntent)
+                true
+            }
+            R.id.action_sponsor -> {
+                val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(Constants.SPONSOR_URL))
+                startActivity(browserIntent)
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
         }
     }
 }
