+++ app/src/main/java/org/musicbrainz/picard/barcodescanner/activities/PreferencesActivity.kt
@@ -20,30 +20,33 @@
 package org.musicbrainz.picard.barcodescanner.activities
 
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.*
+import androidx.appcompat.app.AppCompatActivity
 import org.musicbrainz.picard.barcodescanner.R
 import org.musicbrainz.picard.barcodescanner.databinding.ActivityPreferencesBinding
 import org.musicbrainz.picard.barcodescanner.util.Constants
+import org.musicbrainz.picard.barcodescanner.util.Preferences
 import org.musicbrainz.picard.barcodescanner.views.ConnectionStatusView
 
-class PreferencesActivity : BaseActivity() {
+class PreferencesActivity : AppCompatActivity() {
     private var barcode: String? = null
-    private var connectionBox: ConnectionStatusView? = null
     private lateinit var binding: ActivityPreferencesBinding
+    private var mPreferences: Preferences? = null
 
     /** Called when the activity is first created.  */
     public override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityPreferencesBinding.inflate(layoutInflater)
         setContentView(binding.root)
+        supportActionBar!!.setDisplayHomeAsUpEnabled(true)
 
-        connectionBox = findViewById<View>(R.id.connection_status_box) as ConnectionStatusView
-        val actionBar = supportActionBar
-        actionBar!!.setDisplayHomeAsUpEnabled(true)
-        handleIntents()
+        val extras = intent.extras
+        barcode = extras?.getString(Constants.INTENT_EXTRA_BARCODE)
+
         loadFormDataFromPreferences()
         checkConnectButtonEnabled()
         registerEventListeners()
@@ -53,10 +56,17 @@ class PreferencesActivity : BaseActivity() {
         }
     }
 
-    override fun handleIntents() {
-        super.handleIntents()
-        val extras = intent.extras
-        barcode = extras?.getString(Constants.INTENT_EXTRA_BARCODE)
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        // Inflate the menu items for use in the action bar
+        val inflater = menuInflater
+        inflater.inflate(R.menu.main_activity_actions, menu)
+        menu.findItem(R.id.action_settings).isVisible = false
+        return super.onCreateOptionsMenu(menu)
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        onBackPressed()
+        return true
     }
 
     private fun registerEventListeners() {
@@ -111,7 +121,7 @@ class PreferencesActivity : BaseActivity() {
     }
 
     private fun checkConnectionStatus() {
-        connectionBox!!.updateStatus(readIpAddressFromInput(), readPortFromInput())
+        binding.connectionStatusBox.updateStatus(readIpAddressFromInput(), readPortFromInput())
     }
 
     private fun startNextActivity() {
@@ -128,5 +138,31 @@ class PreferencesActivity : BaseActivity() {
         }
         startActivity(intent)
         finish()
+    }
+
+    private val preferences: Preferences
+        get() {
+            if (mPreferences == null) mPreferences = Preferences(this)
+            return mPreferences!!
+        }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle presses on the action bar items
+        return when (item.itemId) {
+            R.id.action_about -> {
+                val aboutIntent = Intent(
+                    this,
+                    AboutActivity::class.java
+                )
+                startActivity(aboutIntent)
+                true
+            }
+            R.id.action_sponsor -> {
+                val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(Constants.SPONSOR_URL))
+                startActivity(browserIntent)
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
     }
 }
